move app/options into app/views/admin
